- Update Nuget packages, - Update Browser Project and make wasm work. - Overall cleanup

+++ Forge.Forms/src/Forge.Forms.Avalonia/Controls/TextProperties.cs
@@ -5,14 +5,14 @@ namespace Forge.Forms.AvaloniaUI.Controls;
 public static class TextProperties
 {
     public static readonly AttachedProperty<double> TitleFontSizeProperty =
-        AvaloniaProperty.RegisterAttached<AvaloniaObject,double>(
+        AvaloniaProperty.RegisterAttached<AvaloniaObject, double>(
             "TitleFontSize",
             typeof(TextProperties),
             20d,
             true);
 
     public static readonly AttachedProperty<double> HeadingFontSizeProperty =
-        AvaloniaProperty.RegisterAttached<AvaloniaObject,double>(
+        AvaloniaProperty.RegisterAttached<AvaloniaObject, double>(
             "HeadingFontSize",
             typeof(TextProperties),
             15d,
@@ -20,15 +20,15 @@ public static class TextProperties
 
 
     public static readonly AttachedProperty<double> TextFontSizeProperty =
-        AvaloniaProperty.RegisterAttached<AvaloniaObject,double>(
+        AvaloniaProperty.RegisterAttached<AvaloniaObject, double>(
             "TextFontSize",
             typeof(TextProperties),
             13d,
             true);
-    
+
     public static double GetTitleFontSize(AvaloniaObject element)
     {
-        return (double)element.GetValue(TitleFontSizeProperty);
+        return element.GetValue(TitleFontSizeProperty);
     }
 
     public static void SetTitleFontSize(AvaloniaObject element, double value)
@@ -38,7 +38,7 @@ public static class TextProperties
 
     public static double GetHeadingFontSize(AvaloniaObject element)
     {
-        return (double)element.GetValue(HeadingFontSizeProperty);
+        return element.GetValue(HeadingFontSizeProperty);
     }
 
     public static void SetHeadingFontSize(AvaloniaObject element, double value)
@@ -48,7 +48,7 @@ public static class TextProperties
 
     public static double GetTextFontSize(AvaloniaObject element)
     {
-        return (double)element.GetValue(TextFontSizeProperty);
+        return element.GetValue(TextFontSizeProperty);
     }
 
     public static void SetTextFontSize(AvaloniaObject element, double value)
